requirements.txt
torch==1.6.0
torchvision==0.7.0
numpy==1.19.2
Pillow==7.0.0
json5==0.9.5


